docs(dep-006): fix formatting
The formatting on <local-port>:<remote-port> is being interpreted as an
HTML-like tag in GitHub's markdown renderer. This changes it to all caps
like an env var.

I also tried to make it slightly clearer that the first port was the one
on localhost, and the second was the one for the container. Earlier, I
got confused and interpreted "local" as "the port local to the binary"
and "remote" as "the tunnel".